requirements.txt
# Python dependencies for Vercel serverless functions and RAG
# These will be automatically installed when deploying to Vercel

# Core dependencies
Flask==3.1.0
flask-cors==6.0.1

# LangChain and AI
langchain==0.3.13
langchain-openai==0.2.12
langchain-core==0.3.28
langchain-community==0.3.13

# OpenAI
openai==1.59.5

# Vector stores and embeddings
faiss-cpu==1.9.0
tiktoken==0.8.0

# Document loaders and processing
beautifulsoup4==4.12.3
lxml==5.3.0

# LangSmith tracing (optional)
langsmith==0.2.11

# Other dependencies
python-dotenv==1.0.1

